# swagger.yml
openapi: 3.0.0
info:
  title: API de Mensagens
  description: API para gerenciar mensagens
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local

paths:
  /api/data:
    get:
      summary: Lista todas as mensagens ou filtra por texto
      parameters:
        - in: query
          name: message
          schema:
            type: string
          description: Texto para filtrar mensagens
      responses:
        '200':
          description: Lista de mensagens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mensagem'

    post:
      summary: Cria uma nova mensagem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MensagemInput'
      responses:
        '201':
          description: Mensagem criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/data/{id}:
    get:
      summary: Obtém uma mensagem específica
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Mensagem encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'
        '404':
          description: Mensagem não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Atualiza completamente uma mensagem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MensagemInput'
      responses:
        '200':
          description: Mensagem atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Mensagem não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Atualiza parcialmente uma mensagem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MensagemInput'
      responses:
        '200':
          description: Mensagem modificada parcialmente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Mensagem não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Remove uma mensagem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Mensagem removida com sucesso
        '404':
          description: Mensagem não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Mensagem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string

    MensagemInput:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string